Fix: Properly check the user inputted Port Number within the Create Instance dialog. Fix: Properly store the server's Private Port Number as an unsigned short within the ServerInfo class. Fix: When using Port Number "0" within the Create Instance Dialog we will now properly show the User our randomly generated port number by updating the UI.

+++ src/serverinfo.cpp
@@ -628,12 +628,12 @@ void ServerInfo::setPublicIP(const QString& value)
     publicIP = value;
 }
 
-int ServerInfo::getPrivatePort() const
+quint16 ServerInfo::getPrivatePort() const
 {
     return privatePort;
 }
 
-void ServerInfo::setPrivatePort(int value)
+void ServerInfo::setPrivatePort(quint16 value)
 {
     privatePort = value;
 }
